src/requirements.txt
# Allow more static typing
typing==3.6.2
# Easily send HTTP requests
requests==2.11.1
# Easily generate podcast feeds
podgen==1.0.0
# Get duration of an MP3 file
tinytag==0.16.0
# Easily cache properties which could take some time to calculate
cached-property==1.3.0
# Make it possible to work with timezones
pytz==2016.6.1
# Testing without boilerplate
pytest==2.9.2
# For Chimera episode source
Markdown==2.6.8
# For progress bar
clint==0.5.1
# For parsing dates from RadioRevolt.no
tonyg-rfc3339==0.1
# Simple webserver framework
Flask==0.11.1
# Our own webserver
uWSGI==2.0.15
# Image manipulation
Pillow==3.3.0
# Database connection (for SlugList)
psycopg2==2.6.2
PyYAML==3.12
## The following requirements were added by pip freeze:
args==0.1.0
click==6.6
dateutils==0.6.6
feedgen==0.3.2
future==0.16.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.7.3
MarkupSafe==0.23
py==1.4.33
python-dateutil==2.6.0
six==1.10.0
Werkzeug==0.11.10
